build: remove storage package
Storage is tightly coupled to the web application, hence it
does not make sense to have it as an independent package.
