+++ apps/web/package.json
@@ -17,12 +17,12 @@
     "@blueprintjs/icons": "^5.9.0",
     "@blueprintjs/select": "^5.1.5",
     "@repo/github": "workspace:*",
-    "@repo/storage": "workspace:*",
     "@repo/types": "workspace:*",
     "@tanstack/react-query": "^4.32.0",
     "@tanstack/react-query-devtools": "^4.32.1",
     "clsx": "^2.1.1",
     "comlink": "^4.4.1",
+    "dexie": "^4.0.7",
     "dexie-react-hooks": "^1.1.7",
     "javascript-time-ago": "^2.5.10",
     "normalize.css": "^8.0.1",
