+++ apps/web/src/components/CommandBar.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, test, beforeAll, afterAll } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { userEvent, type UserEvent } from "@testing-library/user-event";
-import { db } from "@repo/storage";
+import { db } from "../lib/db";
 import CommandBar from "./CommandBar";
 import { HotkeysProvider } from "@blueprintjs/core";
 import { mockPull } from "@repo/testing";
