+++ apps/web/src/components/CommandBar.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState } from "react";
 import { HotkeysTarget2, HotkeyConfig, MenuItem } from "@blueprintjs/core";
 import { ItemPredicate, ItemRenderer, Omnibar } from "@blueprintjs/select";
 import type { Pull } from "@repo/types";
-import { usePulls } from "../db";
+import { usePulls } from "../lib/queries";
 
 type Props = {
     className?: string,
