+++ apps/web/src/worker/instance.ts
@@ -1,6 +1,6 @@
 import * as Comlink from "comlink";
 import { groupBy, unique } from "remeda";
-import { db } from "@repo/storage";
+import { db } from "../lib/db";
 import { splitQueries, type Pull, type Connection } from "@repo/types";
 import { GitHubClient, isInAttentionSet } from "@repo/github";
 import { gitHubClient } from "../github";
