--- packages/storage/.eslintrc.cjs
@@ -1,10 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  root: true,
-  ignorePatterns: [".eslintrc.cjs"],
-  extends: ["@repo/eslint-config/index.js"],
-  parser: "@typescript-eslint/parser",
-  parserOptions: {
-    project: true,
-  },
-};
