--- packages/storage/README.md
@@ -1,3 +1,0 @@
-# packages/storage
-
-Contains the definition of the database. Data is stored in IndexedDB using [Dexie.js](https://dexie.org).
