--- packages/storage/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@repo/storage",
-  "version": "0.0.0",
-  "exports": {
-    ".": {
-      "types": "./build/index.d.ts",
-      "default": "./build/index.js"
-    }
-  },
-  "license": "MIT",
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "lint": "eslint \"**/*.ts\"",
-    "test": "vitest run",
-    "coverage": "vitest run --coverage"
-  },
-  "devDependencies": {
-    "@repo/eslint-config": "workspace:*",
-    "@repo/testing": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
-    "@vitest/coverage-istanbul": "^2.0.5",
-    "eslint": "^8.57.0",
-    "fake-indexeddb": "^6.0.0",
-    "typescript": "^5.5.4",
-    "vitest": "^2.0.5"
-  },
-  "dependencies": {
-    "@repo/types": "workspace:*",
-    "dexie": "^4.0.7"
-  }
-}
