--- packages/storage/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@repo/typescript-config/base.json",
-  "compilerOptions": {
-    "lib": ["es2015"],
-    "outDir": "build",
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"],
-}
