--- packages/storage/vitest.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    setupFiles: ["fake-indexeddb/auto"],
-    coverage: {
-      include: ["src/**"],
-      provider: "istanbul",
-    },
-  },
-});
